src/research/requirements.txt
auto_mix_prep
beautifulsoup4
fitz
pikepdf
pypdf
python_magic
Requests
rich
selenium==4.35.0
pytest==8.4.1
trio==0.30.0
pytest-trio==0.8.0
pytest-rerunfailures==15.1
flake8==7.3.0
requests==2.32.5
tox==4.28.4
pytest-xdist==3.8.0
textual
typer
web-browser
argparse

